+++ module/canvas/index.ts
@@ -1,15 +1,20 @@
+import { GurpsModule } from 'module/gurps-module.js'
 import { GurpsRulerV12 } from './ruler-12.js'
-import { GurpsRuler } from './ruler.js'
+import { registerRuler } from './ruler.js'
 
-export function init() {
+function init() {
+  console.log('GURPS | Initializing GURPS Canvas Module')
+
   Hooks.once('init', () => {
-    console.log('GURPS | Initializing GURPS Canvas Module')
-
     if (!game.release) return
     if (game.release?.generation >= 13) {
-      CONFIG.Canvas.rulerClass = GurpsRuler
+      registerRuler()
     } else {
       CONFIG.Canvas.rulerClass = GurpsRulerV12
     }
   })
+}
+
+export const Canvas: GurpsModule = {
+  init,
 }