+++ module/combat/combat.ts
@@ -9,7 +9,7 @@ class GurpsCombat<SubType extends Combat.SubType = Combat.SubType> extends Comba
       for (const combatant of this.combatants) {
         const tokenId = combatant.token?.id ?? null
         if (tokenId === null) continue
-        const token = canvas.tokens?.get(tokenId)
+        const token = canvas?.tokens?.get(tokenId)
         if (token) {
           await token.removeManeuver()
         }