+++ module/token/index.ts
@@ -1,12 +1,28 @@
-import GurpsTokenHUD from './hud.js'
+import { GurpsModule } from 'module/gurps-module.js'
 import { GurpsToken } from './gurps-token.js'
+import { GurpsTokenHUD } from './token-hud-12.js'
+import { registerTokenHUD } from './token-hud.js'
+import { registerTokenRuler } from './token-ruler.js'
 
 export * from './gurps-token.js'
-export * from './hud.js'
+export * from './quick-roll-settings.js'
 
-export function init(): void {
+function init(): void {
+  console.log('GURPS | Initializing GURPS Token module.')
   Hooks.once('init', () => {
     CONFIG.Token.objectClass = GurpsToken
-    CONFIG.Token.hudClass = GurpsTokenHUD
+
+    if (!game.release) return
+    if (game.release?.generation >= 13) {
+      // COMPATIBILITY: v12
+      registerTokenHUD()
+      registerTokenRuler()
+    } else {
+      CONFIG.Token.hudClass = GurpsTokenHUD
+    }
   })
+}
+
+export const Token: GurpsModule = {
+  init,
 }